+++ src/TextShow/Data/Binary.hs
@@ -16,7 +16,9 @@ Portability: GHC
 module TextShow.Data.Binary () where
 
 import Data.Binary.Get.Internal (Decoder(..))
-import Data.Monoid.Compat
+
+import Prelude ()
+import Prelude.Compat
 
 import TextShow (TextShow(..), TextShow1(..), Builder, fromString, showbPrec1)